funcat: update README for further plan
Due to hands-on practice on Alibaba Cloud, currently I recommend
python3.6.8 for this repo. But the python EOL is the problem that
we have to face up in the further.

So, I update the README to show the plan to support higher python
version. Furthmore, container deployment (such as docker.eg.) will
become a optional.

Signed-off-by: Guanjun <guanjun@linux.alibaba.com>